+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Change log
 -----------
 
+* Change partition type of resin-conf from vfat to ext4 [Theodor]
+* Move config.json to resin-boot partition [Theodor]
+* Change resin-boot mount point from /boot to /mnt/boot [Theodor]
 * Make dropbear host key generation atomic [Michal]
 * Change the TUN device used by OpenVPN to resin-vpn to minimize conflict with user applications using tun0 [Praneeth]
 * Update supervisor to v2.1.0 [Pablo]